Move IEEE 802.11 structs definitions in another header.
opendroneid.h header is the public API of libopendroneid and IEEE 802.11
structs are internal structure.
